requirements.txt
attrs
numpy
torch
torchvision
Pillow
tqdm